time: widen tm_mday before subtracting, in mktime and timegm
`days = __days_from_civil(y, m, 1) + (tm->tm_mday - 1)` evaluates the
subtraction in `int`.  At tm_mday == INT_MIN that is signed overflow --
undefined behaviour inside the one function whose contract is that
out-of-range fields are accepted and normalized rather than rejected.

The comment three lines below already said what to do, and said it about
the neighbouring fields only: "tm_hour/tm_min are `int`, unbounded like
tm_mon above ...; widen before multiplying so a large caller-supplied
value overflows in `long long`, not in `int`".  tm_mday needed the same
widening and did not have it.

Found by fuzz/fuzz_time.c on its first ninety-second run.  UBSan:
"signed integer overflow: -2147483648 - 1 cannot be represented in type
'int'" at src/time/timegm.c:23.  The reproducing input no longer trips
it after this change; the pre-change binary is what produced it.

fuzz_time.c's own fold() had the mirror-image defect -- hi - lo computed
in `int` for the +-2e9 time_t range -- reported by UBSan on the empty
input, before the harness had run on anything else.  Fixed here too.

+++ fuzz/fuzz_time.c
@@ -115,13 +115,16 @@ static long long rd64(const unsigned char *p)
 	return (long long)u;
 }
 
-/* Fold v into [lo, hi] without a modulus that could be by zero, and
- * without depending on C's sign rules for %: the span is at most a few
- * thousand here, so an unsigned remainder is exact and total. */
+/* Fold v into [lo, hi].  The width is computed in `unsigned`, not in
+ * `int`: the widest range folded to here is the +-2e9 time_t one, and
+ * hi - lo for that is 4e9, which overflows `int`.  UBSan reported it on
+ * the very first run of this harness, on the empty input -- the harness
+ * had the defect, not the library, which is the right way round for it
+ * to be found. */
 static int fold(int v, int lo, int hi)
 {
-	unsigned span = (unsigned)(hi - lo) + 1u;
-	return lo + (int)((unsigned)v % span);
+	unsigned span = ((unsigned)hi - (unsigned)lo) + 1u;
+	return (int)((unsigned)lo + (unsigned)v % span);
 }
 
 /* The harness's own decimal formatter.  Not ntlibc's snprintf, which is
